Move prepareTree to its own module
