+++ src/EventHandling.hs
@@ -95,60 +95,3 @@ runEvent fromEvent@FromEvent { message, location } component = case component of
   Hide x -> runEvent fromEvent x
 
   x -> pure []
-
-{-|
-
-This walks through the tree and collects actions that should be run
-only once, and sets their run value to True.  It's up to something
-else to actually send the actions.
-
-It also assigns a location to message and effect handlers.
-
--}
-
-prepareGraph :: Purview a m -> (Purview a m, [FromEvent])
-prepareGraph = prepareGraph' [] []
-
-type Location = [Int]
-
-prepareGraph' :: Location -> Location -> Purview a m -> (Purview a m, [FromEvent])
-prepareGraph' parentLocation location component = case component of
-  Attribute attrs cont ->
-    let result = prepareGraph' parentLocation location cont
-    in (Attribute attrs (fst result), snd result)
-
-  Html kind children ->
-    let result = fmap (\(index, child) -> prepareGraph' parentLocation (index:location) child) (zip [0..] children)
-    in (Html kind (fmap fst result), concatMap snd result)
-
-  EffectHandler _ploc _loc state handler cont ->
-    let
-      rest = fmap (prepareGraph' location (0:location)) cont
-    in
-      ( EffectHandler (Just parentLocation) (Just location) state handler (\state' -> fst (rest state'))
-      , snd (rest state)
-      )
-
-  Once effect hasRun cont ->
-    let send message =
-          FromEvent
-            { event = "once"
-            , message = toJSON message
-            , location = Just location
-            }
-    in if not hasRun then
-        let
-          rest = prepareGraph' parentLocation location cont
-        in
-          (Once effect True (fst rest), [effect send] <> (snd rest))
-       else
-        let
-          rest = prepareGraph' parentLocation location cont
-        in
-          (Once effect True (fst rest), snd rest)
-
-  Hide x ->
-    let (child, actions) = prepareGraph' parentLocation location x
-    in (Hide child, actions)
-
-  component' -> (component', [])
