+++ src/EventHandlingSpec.hs
@@ -14,9 +14,10 @@ import Data.Aeson.TH
 import Data.Time
 
 import Component
-import Rendering
-import Events
 import EventHandling
+import Events
+import PrepareTree
+import Rendering
 
 
 data TestAction = Up | Down
@@ -169,7 +170,7 @@ spec = parallel $ do
       chan <- newTChanIO
 
       let
-        locatedGraph = fst $ prepareGraph component
+        locatedGraph = fst $ prepareTree component
         event1 = FromEvent { event="click", message=toJSON Up, location=Just [1, 0] }
 
       afterEvent1 <- apply chan event1 locatedGraph
@@ -204,7 +205,7 @@ spec = parallel $ do
 
         chan <- newTChanIO
 
-        let locatedGraph = fst $ prepareGraph handler
+        let locatedGraph = fst $ prepareTree handler
 
         render locatedGraph `shouldBe` "<div handler=\"[]\"><div><div handler=\"[1,0]\">0</div></div></div>"
 
@@ -242,7 +243,7 @@ spec = parallel $ do
 
         chan <- newTChanIO
 
-        let locatedGraph = fst $ prepareGraph handler
+        let locatedGraph = fst $ prepareTree handler
 
         render locatedGraph `shouldBe` "<div handler=\"[]\"><div><div handler=\"[1,0]\">0</div></div></div>"
 
@@ -256,121 +257,6 @@ spec = parallel $ do
         receivedEvent2 <- atomically $ readTChan chan
         -- correctly targeted to self
         receivedEvent2 `shouldBe` FromEvent {event = "internal", message = String "Down", location = Just [1,0]}
-
-  describe "prepareGraph" $ do
-
-    it "sets hasRun to True" $ do
-      let
-        display time = div
-          [ text (show time)
-          , onClick ("setTime" :: String) $ div [ text "check time" ]
-          ]
-
-        startClock cont state = Once (\send -> send ("setTime" :: String)) False (cont state)
-
-        timeHandler = EffectHandler Nothing Nothing Nothing handle
-
-        handle :: String -> Maybe UTCTime -> IO (Maybe UTCTime, [DirectedEvent String String])
-        handle "setTime" _     = do
-          time <- getCurrentTime
-          pure (Just time, [])
-        handle _         state = pure (state, [])
-
-        component = timeHandler (startClock display)
-
-      show (fst (prepareGraph component))
-        `shouldBe`
-        "EffectHandler Just [] Just [] Once True div [  \"Nothing\" Attr div [  \"check time\" ]  ] "
-
-      length (snd (prepareGraph component))
-        `shouldBe`
-        1
-
-    it "stops collecting the action if it has already run" $ do
-      let
-        display time = div
-          [ text (show time)
-          , onClick ("setTime" :: String) $ div [ text "check time" ]
-          ]
-
-        startClock cont state = Once (\send -> send ("setTime" :: String)) False (cont state)
-
-        timeHandler = EffectHandler Nothing Nothing Nothing handle
-
-        handle :: String -> Maybe UTCTime -> IO (Maybe UTCTime, [DirectedEvent String String])
-        handle "setTime" _     = do
-          time <- getCurrentTime
-          pure (Just time, [])
-        handle _         state = pure (state, [])
-
-        component = timeHandler (startClock display)
-
-      let
-        run1 = prepareGraph component
-        run2 = prepareGraph (fst run1)
-        run3 = prepareGraph (fst run2)
-
-      length (snd run1) `shouldBe` 1
-      length (snd run2) `shouldBe` 0
-      length (snd run3) `shouldBe` 0  -- for a bug where it was resetting run
-
-    it "assigns a location to handlers" $ do
-      let
-        timeHandler = effectHandler Nothing handle
-
-        handle :: String -> Maybe UTCTime -> IO (Maybe UTCTime, [DirectedEvent String String])
-        handle "setTime" _     = do
-          time <- getCurrentTime
-          pure (Just time, [])
-        handle _         state = pure (state, [])
-
-        component = timeHandler (const (Text ""))
-
-      component `shouldBe` Hide (EffectHandler Nothing Nothing Nothing handle (const (Text "")))
-
-      let
-        graphWithLocation = fst (prepareGraph component)
-
-      graphWithLocation `shouldBe` Hide (EffectHandler (Just []) (Just []) Nothing handle (const (Text "")))
-
-    it "assigns a different location to child handlers" $ do
-      let
-        timeHandler = effectHandler Nothing handle
-
-        handle :: String -> Maybe UTCTime -> IO (Maybe UTCTime, [DirectedEvent String String])
-        handle "setTime" _     = do
-          time <- getCurrentTime
-          pure (Just time, [])
-        handle _         state = pure (state, [])
-
-        component = div
-          [ timeHandler (const (Text ""))
-          , timeHandler (const (Text ""))
-          ]
-
-        graphWithLocation = fst (prepareGraph component)
-
-      show graphWithLocation
-        `shouldBe`
-        "div [  Hide EffectHandler Just [] Just [0] \"\" Hide EffectHandler Just [] Just [1] \"\" ] "
-
-    it "assigns a different location to nested handlers" $ do
-      let
-        timeHandler = effectHandler Nothing handle
-
-        handle :: String -> Maybe UTCTime -> IO (Maybe UTCTime, [DirectedEvent String String])
-        handle "setTime" _     = do
-          time <- getCurrentTime
-          pure (Just time, [])
-        handle _         state = pure (state, [])
-
-        component =
-          timeHandler (const (timeHandler (const (Text ""))))
-
-
-        graphWithLocation = fst (prepareGraph component)
-
-      show graphWithLocation `shouldBe` "Hide EffectHandler Just [] Just [] Hide EffectHandler Just [] Just [0] \"\""
 
 
 main :: IO ()
